refactor(console): split store
