+++ src/console/src/store/mod.rs
@@ -1,3 +1,5 @@
 pub mod heap;
 mod services;
-pub mod stable;
+mod stable;
+
+pub use stable::*;
