+++ src/console/src/store/stable/payments.rs
@@ -1,0 +1,124 @@
+use crate::store::services::{mutate_stable_state, read_stable_state};
+use crate::types::ledger::{Payment, PaymentStatus};
+use crate::types::state::{Payments, PaymentsStable, StableState};
+use ic_cdk::api::time;
+use ic_ledger_types::BlockIndex;
+use junobuild_shared::structures::collect_stable_map_from;
+use junobuild_shared::types::state::UserId;
+
+pub fn is_known_payment(block_index: &BlockIndex) -> bool {
+    read_stable_state(|stable| stable.payments.contains_key(block_index))
+}
+
+pub fn insert_new_payment(
+    user: &UserId,
+    block_index: &BlockIndex,
+) -> Result<Payment, &'static str> {
+    mutate_stable_state(|stable| insert_new_payment_impl(user, block_index, stable))
+}
+
+fn insert_new_payment_impl(
+    user: &UserId,
+    block_index: &BlockIndex,
+    state: &mut StableState,
+) -> Result<Payment, &'static str> {
+    let existing_mission_control = state.mission_controls.get(user);
+
+    match existing_mission_control {
+        None => Err("User does not have a mission control center"),
+        Some(mission_control) => {
+            let now = time();
+
+            let new_payment = Payment {
+                mission_control_id: mission_control.mission_control_id,
+                block_index_payment: *block_index,
+                block_index_refunded: None,
+                status: PaymentStatus::Acknowledged,
+                created_at: now,
+                updated_at: now,
+            };
+
+            state.payments.insert(*block_index, new_payment.clone());
+
+            Ok(new_payment)
+        }
+    }
+}
+
+pub fn update_payment_completed(block_index: &BlockIndex) -> Result<Payment, &'static str> {
+    mutate_stable_state(|stable| update_payment_completed_impl(block_index, stable))
+}
+
+fn update_payment_completed_impl(
+    block_index: &BlockIndex,
+    state: &mut StableState,
+) -> Result<Payment, &'static str> {
+    let payment = state.payments.get(block_index);
+
+    match payment {
+        None => Err("Payment not found."),
+        Some(payment) => {
+            let now = time();
+
+            let updated_payment = Payment {
+                mission_control_id: payment.mission_control_id,
+                block_index_payment: payment.block_index_payment,
+                block_index_refunded: Some(payment.block_index_payment),
+                status: PaymentStatus::Completed,
+                created_at: payment.created_at,
+                updated_at: now,
+            };
+
+            state.payments.insert(*block_index, updated_payment.clone());
+
+            Ok(updated_payment)
+        }
+    }
+}
+
+pub fn update_payment_refunded(
+    block_index_payment: &BlockIndex,
+    block_index_refunded: &BlockIndex,
+) -> Result<Payment, &'static str> {
+    mutate_stable_state(|stable| {
+        update_payment_refunded_impl(block_index_payment, block_index_refunded, stable)
+    })
+}
+
+fn update_payment_refunded_impl(
+    block_index_payment: &BlockIndex,
+    block_index_refunded: &BlockIndex,
+    state: &mut StableState,
+) -> Result<Payment, &'static str> {
+    let payment = state.payments.get(block_index_payment);
+
+    match payment {
+        None => Err("Payment to refund not found."),
+        Some(payment) => {
+            let now = time();
+
+            let updated_payment = Payment {
+                mission_control_id: payment.mission_control_id,
+                block_index_payment: payment.block_index_payment,
+                block_index_refunded: Some(*block_index_refunded),
+                status: PaymentStatus::Refunded,
+                created_at: payment.created_at,
+                updated_at: now,
+            };
+
+            state
+                .payments
+                .insert(*block_index_payment, updated_payment.clone());
+
+            Ok(updated_payment)
+        }
+    }
+}
+
+pub fn list_payments() -> Payments {
+    read_stable_state(|stable| list_payments_impl(&stable.payments))
+}
+
+fn list_payments_impl(payments: &PaymentsStable) -> Payments {
+    collect_stable_map_from(payments)
+}
